Refactor!
