+++ src/common/hooks/externalClickHandler.ts
