+++ src/components/home/pages/info/index.tsx
@@ -1,6 +1,6 @@
-import { IconProps } from "@lib/types";
 import React, { useEffect, useState } from "react";
-import profile_image from "../../../IMG/profile-sm.webp";
+import profile_image from "src/common/images/profile-sm.webp";
+import { IconProps } from "src/common/types";
 
 function Component({ Icon }: IconProps) {
     const [show, setShow] = useState<boolean>(false);
