+++ src/components/index.ts
@@ -1,8 +1,8 @@
 export { default as Sidebar } from "./shared/sidebar";
 export { default as Navbar } from "./shared/navigation";
-export { default as Info } from "./pages/info";
-export { default as Dashboard } from "./pages/dashboard";
-export { default as Routes } from "./routes";
+export { default as Info } from "./home/pages/info";
+export { default as Dashboard } from "./home/pages/dashboard";
+export { default as Routes } from "./home";
 export { default as NewsCarousel } from "./shared/news-reel";
 export { default as Card } from "./shared/card";
 export { default as Footer } from "./shared/footer";
