+++ src/components/shared/footer/index.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { IconProps } from "@lib/types";
+import { IconProps } from "src/common/types";
 
 function Component({ Icon }: IconProps) {
     return (
