+++ tsconfig.json
@@ -15,10 +15,7 @@
         "isolatedModules": true,
         "noEmit": true,
         "jsx": "react-jsx",
-        "baseUrl": "./",
-        "paths": {
-            "@lib/types": ["./lib/types.ts"]
-        }
+        "baseUrl": "./"
     },
     "include": ["src"]
 }
